DBアクセス

+++ src/main/java/net/code/TrSellItemsRepository.java
@@ -2,6 +2,6 @@ package net.code;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrSellItemsRepository extends JpaRepository<TrSellItemsEntity, Long> {
+public interface TrSellItemsRepository extends JpaRepository<TrSellItemsEntity, Integer> {
 
 }
